pyproject.toml
[tool.poetry]
name = "startup-boilerplate"
version = "0.1.0"
description = "Boilerplate to create fun projects"
authors = ["Maksadbek Akhmedov <a.maksadbek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"
flask-jwt-extended = "^4.6.0"
flask-login = "^0.6.3"
flask-wtf = "^1.2.1"
email-validator = "^2.1.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
